+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 ## [Unreleased]
 ### Added
 
+- assigning `Sycamore::Nothing` via `Tree#[]=` removes a child tree
 - `Tree#search` for searching the tree for one or multiple nodes or a tree
 - `Tree#node!` as a more strict variant of `Tree#node`, which raises an error 
   when no node present